Updated version and added an event forwarder class
In IceCoffee.Common.csproj , the version number was updated from 1.3.2 to 1.3.3, and the versions of Microsoft.Extensions.ObjectPool and System.Threading.Channels were updated. Additionally, a reference to System.Reflection.Emit.Lightweight was added for the netstandard2.0 framework.

In EventForwarder.cs , a new EventForwarder class was added to dynamically subscribe to events and call the OnEventRaised method when an event is triggered. This class obtains event information through reflection and uses a dynamic method to create an event handler to correctly handle event parameters.